+++ example/example/settings.py
@@ -63,3 +63,6 @@ TEMPLATES = [{
         ],
     },
 }]
+
+
+TIME_ZONE = 'Europe/London'