+++ tests/models.py
@@ -1,10 +1,20 @@
-from conman.routes.handlers import URLConfHandler
+from django.db import models
+
+from conman.routes.handlers import TemplateHandler, URLConfHandler, ViewHandler
 from conman.routes.managers import RouteManager
 from conman.routes.models import Route
 
 
 class RouteSubclass(Route):
     """A Route for testing direct subclasses of Route."""
+    template_name = 'ignored.html'  # Added to keep checks happy.
+    # Silence RemovedInDjango20Warning about manager inheritance.
+    objects = RouteManager()
+
+
+class ViewRoute(Route):
+    """A Route for testing ViewHandler."""
+    handler_class = ViewHandler
     # Silence RemovedInDjango20Warning about manager inheritance.
     objects = RouteManager()
 
@@ -16,6 +26,15 @@ class RouteSubclass(Route):
 
 class NestedRouteSubclass(RouteSubclass):
     """A Route for testing nested subclasses of Route."""
+    # Silence RemovedInDjango20Warning about manager inheritance.
+    objects = RouteManager()
+
+
+class TemplateRoute(Route):
+    """A Route for testing TemplateHandler."""
+    content = models.TextField()
+    handler_class = TemplateHandler
+    template_name = 'basic_template.html'
     # Silence RemovedInDjango20Warning about manager inheritance.
     objects = RouteManager()